Develop stream 2025-01-23 (#678)
* Avoid evaluating CMake LINK_LANGUAGE generator expressions in CMake <3.18 (#667)

Recently, we added CMake HIP Language support. That change introduced generator expressions
that use the LINK_LANGUAGE keyword. This keyword is only supported in CMake 3.18+.
The minimum CMake version we require is 3.16.

As a result, if you built with CMake 3.16 - 3.18, you'd get a build failure.

This change adds checks around all the spots where the LINK_LANGUAGE keyword is used
to avoid this problem.

It also adds a description of the CMake HIP Language support to the changelog (previously overlooked).

* Fix Windows CMake link issue (#675)

A recent cmake change added a condition check that avoids evaluating
a LINK_LANGUAGE generator expression if HIP_USECXX is OFF (since
LINK_LANGUAGE is unsupported in CMake < 3.18). However, it didn't
account for the fact that the LINK_LANGUAGE expression is embedded in a
conditional generator expression.

This change adds an "else" block that ensures the "false" side of the
conditional generator expression is still evaluated when the LINK_LANGUAGE
code is skipped. This fixes some build issues on Windows.

Also updates the readme to include recent changes to the way projects
must link to rocPRIM.

* Resolve "Add hipGraph compatibility for device run length encode"

* Resolve "Investigate RLE (non trivial runs) failure"

* Resolve "Config tuning and dynamic dispatch for device run length encode"

* Resolve "Implement type traits interface for rocPRIM's compiler-extended types and third-party's types"

* remove ordered block id from run length encode

A missing call to __syncthreads() caused this algorithm to occasionally
produce an invalid index and cause a crash. Instead of just adding that
call, we can remove the use of ordered block id by the regular block id,
which is already ordered per HSA guarantees.

There still remains an issue with a hang, but this seems a runtime issue
and will be investigated separately.

* MI200 & MI300 tuning

* Optimizations for MI300 unique performance

* improve type traits interface docs

* Fix static asserts and remove template specialization

* Fix tests triggering static asserts

* Fix benchmarks triggering static asserts

* Set explicit return types

* Resolve "Investigate RLE (non trivial runs) failure"

* Revert "Fix benchmarks triggering static asserts"

This reverts commit a1634a9382e268285c1e21c0743b6a98e67a69ef.

* Revert "Fix tests triggering static asserts"

This reverts commit 24a36cb86283098036a5931562d0b4d3b42ca675.

* Revert "Fix static asserts and remove template specialization"

This reverts commit b2089eeedec6b32415f27d0e3c0860442b0ba5e0.

* Fix formatting and copyright date

* Temp fix: histogram dynamic memory failure on windows

* add ROCPRIM_ENABLE_ASSERTS to cmake to enable asserts in release build
set ROCPRIM_ENABLE_ASSERTS for CI tests

* Add missing docs for `block_run_length_decode`

* Resolve "Add documentation and test for out-of-range behaviour on merge path search"

add test for out of bounds diagonal test for rocprim::merge_path_search
document behaviour for out of bounds diagonal test for rocprim::merge_path_search

* Resolve "Replace __launch_bounds__ with ROCPRIM_LAUNCH_BOUNDS in rocPRIM API"

* Update rmake.py copyright

* fix double buffers in device radix sort

* Resolve "Replace hipMalloc, hipMallocHelper, and hipFree in tests with device_ptr"

* fix device adjacent difference tuning benchmark

* Fix warp dpp reduction

* Fix static assertion for int128 in ONNX Runtime with GNU extensions

* fix GenerateResourceSpec for gfx90a

* Resolve "ROCPRIM_TARGET_ARCH deprecation"

* Fix warnings when generating sphinx docs

* Performance improvements with shuffled warp exchange

* log more information during configure

* Update documentation generation instructions

* Resolve "Replace hipMalloc, hipMallocHelper, and hipFree in benchmarks and tests with device_ptr" Part 4

* Resolve "search_n optimization"

* CI changes to speedup development

* Remove select_arch and select_arch_case

* fix missing equality operator and function constness

* Print arguments and timing info in `batch_memcpy_func` when `debug_synchronous` is true

* Resolve "Replace hipMalloc, hipMallocHelper, and hipFree in benchmarks and tests with device_ptr" Part 3

* Resolve "Replace hipMalloc, hipMallocHelper, and hipFree in benchmarks and tests with device_ptr" Part 1

* Resolve "Replace hipMalloc, hipMallocHelper, and hipFree in benchmarks and tests with device_ptr" Part 2

* Fix rules overwrite for other then MR pipelines

* Resolve "Add stream, memset, array, and unique_ptr support to device_ptr"

* Resolve "Reduce amount of test runs based on environment variable."

* Resolve "Get clarity on skipping of adjacent difference tests"

* Remove windows CI check

* Resolve "Add UnrolledThreadLoad and UnrolledThreadCopy functions"

* Resolve "General changes CCCL 2.6"

* Reduce CI artifact retention time

* Resolve "Add thread_load_volatile_pointer to rocprim"

* Fix test:install on rocm/dev-ubuntu

* Resolve "Nth element - Investigate smaller index sizes where possible"

* Enable int128 tests on Windows

* Resolve "Benchmark and tuning for 128-bit integers"

* Resolve "Verify sorting on device"

* Resolve "Expose lookback_scan_state::create and lookback_scan_state::get_storage_size as ROCPRIM_HOST_DEVICE."

* add naive in-place api for device merge

* To remove error in CI change it to AMDGPU_TARGETS

* fix device_ptr assign move and void types

* Resolve "Add block- and warp-level inclusive_scan with initial value support"

* Fix C++ includes showing up in the docs

* Resolve "Performance regressions during tuning: Adjacent find"

* CI skip label fix

* fix doxygen warning in merge.rst

* Resolve "Autotuning for segmented reduce"

* Resolve "Update thread_store for parity"

* Resolve "Fix compile errors in "rocprim-release-tune" mode"

* Documentation to CI upload

* Resolve "Remove ASM from thread_load and thread_store"

* Add documentation steps to developer guidelines

* Resolve "Refactor lookback_scan_state API"

* Adopt symbol visibility rules in rocPRIM

* remove nm debug messages from cmake

* Remove extra dashes in the documentation its \param and \tparam

* update changelog

* fix(test_device_merge_inplace.cpp): prevent merge inplace tests to tests to grow too big

* Fix test_linking: install the test's libraries too

* chore(rmake.py): bump copyright

* fix(example_type_traits_interface.cpp): explicit return from main

* Fix unintentionally exported shared libs of test_linking

rocm_install adds all libraries to ${PROJECT_NAME_LOWER}-targets
(i.e. rocprim-targets.cmake) unless a custom export file is specified.

---------


[ROCm/rocPRIM commit: 678701d13b9c1464d59843ead73816acb1716008]

+++ rocprim/include/rocprim/device/device_nth_element.hpp
@@ -77,7 +77,7 @@ hipError_t
     const unsigned int     num_blocks            = ceiling_div(size, num_items_per_block);
 
     key_type*                            tree             = nullptr;
-    size_t*                              buckets          = nullptr;
+    unsigned int*                        buckets          = nullptr;
     n_th_element_iteration_data*         nth_element_data = nullptr;
     bool*                                equality_buckets = nullptr;
     nth_element_onesweep_lookback_state* lookback_states  = nullptr;
@@ -248,14 +248,15 @@ template<class Config = default_config,
          class KeysIterator,
          class BinaryFunction
          = ::rocprim::less<typename std::iterator_traits<KeysIterator>::value_type>>
-ROCPRIM_INLINE hipError_t nth_element(void*          temporary_storage,
-                                      size_t&        storage_size,
-                                      KeysIterator   keys,
-                                      size_t         nth,
-                                      size_t         size,
-                                      BinaryFunction compare_function  = BinaryFunction(),
-                                      hipStream_t    stream            = 0,
-                                      bool           debug_synchronous = false)
+ROCPRIM_INLINE
+hipError_t nth_element(void*          temporary_storage,
+                       size_t&        storage_size,
+                       KeysIterator   keys,
+                       size_t         nth,
+                       size_t         size,
+                       BinaryFunction compare_function  = BinaryFunction(),
+                       hipStream_t    stream            = 0,
+                       bool           debug_synchronous = false)
 {
     return detail::nth_element_impl<Config>(temporary_storage,
                                             storage_size,
@@ -358,15 +359,16 @@ template<class Config = default_config,
          class KeysOutputIterator,
          class BinaryFunction
          = ::rocprim::less<typename std::iterator_traits<KeysInputIterator>::value_type>>
-ROCPRIM_INLINE hipError_t nth_element(void*              temporary_storage,
-                                      size_t&            storage_size,
-                                      KeysInputIterator  keys_input,
-                                      KeysOutputIterator keys_output,
-                                      size_t             nth,
-                                      size_t             size,
-                                      BinaryFunction     compare_function  = BinaryFunction(),
-                                      hipStream_t        stream            = 0,
-                                      bool               debug_synchronous = false)
+ROCPRIM_INLINE
+hipError_t nth_element(void*              temporary_storage,
+                       size_t&            storage_size,
+                       KeysInputIterator  keys_input,
+                       KeysOutputIterator keys_output,
+                       size_t             nth,
+                       size_t             size,
+                       BinaryFunction     compare_function  = BinaryFunction(),
+                       hipStream_t        stream            = 0,
+                       bool               debug_synchronous = false)
 {
     using key_type = typename std::iterator_traits<KeysInputIterator>::value_type;
     static_assert(
